moves caching data service to data package
