+++ webapp/src/main/webapp/src/app/assessments/items/item-exemplar/item-exemplar.component.spec.ts
@@ -5,7 +5,7 @@ import { ItemScoringGuideMapper } from "./item-scoring-guide.mapper";
 import { CommonModule } from "../../../shared/common.module";
 import { Observable } from "rxjs";
 import { ItemScoringGuide } from "./model/item-scoring-guide.model";
-import { CachingDataService } from "../../../shared/cachingData.service";
+import { CachingDataService } from "../../../shared/data/caching-data.service";
 import { MockDataService } from "../../../../test/mock.data.service";
 import { DataService } from "../../../shared/data/data.service";
