+++ webapp/src/main/webapp/src/app/school-grade/organization.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from "@angular/core";
 import { UserService } from "../user/user.service";
 import { School } from "../user/model/school.model";
 import { Observable } from "rxjs/Observable";
-import { CachingDataService } from "../shared/cachingData.service";
+import { CachingDataService } from "../shared/data/caching-data.service";
 
 @Injectable()
 export class OrganizationService {
